fix: EMERGENCY - Resolve actor sheet loading by separating Node.js imports - Created theme-config-utils.js for browser-compatible utilities - Updated theme-manager.js to import from browser-safe file - Fixed 'fs' bare specifier error that broke actor sheets - CLI scripts still work via theme-utils.js inheritance - Both v12 and v13 containers tested and working
